feat: initial model function setup
