+++ prisma/schema.prisma
@@ -103,7 +103,7 @@ model apartment_booking {
   room_type      String?         @db.VarChar(255)
   apartment_id   Int?
   users          users           @relation(fields: [user_id], references: [id], onDelete: SetNull, onUpdate: NoAction)
-  apartment      apartment?      @relation(fields: [apartment_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "fk_apartment_id")
+  apartment      apartment?      @relation(fields: [apartment_id], references: [id], onDelete: Cascade, onUpdate: NoAction, map: "fk_apartment_id")
   room           room?           @relation(fields: [room_id], references: [id], onDelete: NoAction, onUpdate: NoAction, map: "fk_room_id")
 }
 
@@ -111,8 +111,8 @@ model apartment_owner {
   id           Int       @id @default(autoincrement())
   user_id      Int
   apartment_id Int
-  apartment    apartment @relation(fields: [apartment_id], references: [id], onDelete: NoAction, onUpdate: NoAction)
-  users        users     @relation(fields: [user_id], references: [id], onDelete: SetNull, onUpdate: NoAction)
+  apartment    apartment @relation(fields: [apartment_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
+  users        users     @relation(fields: [user_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
 }
 
 model apartment_picture {
@@ -120,7 +120,7 @@ model apartment_picture {
   name         String?   @db.VarChar(255)
   file_path    String
   apartment_id Int
-  apartment    apartment @relation(fields: [apartment_id], references: [id], onDelete: NoAction, onUpdate: NoAction)
+  apartment    apartment @relation(fields: [apartment_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
 }
 
 model appointments {
@@ -655,7 +655,7 @@ model rating {
   comment      String?
   created_at   DateTime?  @default(now()) @db.Timestamptz(6)
   apartment_id Int?
-  apartment    apartment? @relation(fields: [apartment_id], references: [id], onUpdate: NoAction)
+  apartment    apartment? @relation(fields: [apartment_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
   users        users      @relation(fields: [user_id], references: [id], onDelete: SetNull, onUpdate: NoAction)
 }
 
@@ -726,7 +726,7 @@ model room {
   price_end         Float?
   apartment_id      Int?
   apartment_booking apartment_booking[]
-  apartment         apartment?          @relation(fields: [apartment_id], references: [id], onDelete: NoAction, onUpdate: NoAction)
+  apartment         apartment?          @relation(fields: [apartment_id], references: [id], onDelete: Cascade, onUpdate: NoAction)
 }
 
 model route_stops {
@@ -1159,11 +1159,11 @@ model staff {
 }
 
 model verification_token {
-  id         String    @id(map: "VerificationToken_pkey") @db.VarChar(255)
   user_id    Int?
   token      String?   @db.VarChar(255)
   created_at DateTime? @default(now()) @db.Timestamptz(6)
   expires_at DateTime? @db.Timestamptz(6)
+  id         Int       @id(map: "id_primary_key") @default(autoincrement())
   users      users?    @relation(fields: [user_id], references: [id], onDelete: Cascade, onUpdate: NoAction, map: "user_id_fkey")
 
   @@index([user_id], map: "fki_user_id_fkey")
